requirements.txt
absl-py==0.8.0
astor==0.8.0
astroid==2.2.5
certifi==2019.6.16
colorama==0.4.1
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
hdbscan==0.8.22
isort==4.3.21
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
llvmlite==0.29.0
Markdown==3.1.1
matplotlib==3.1.1
mccabe==0.6.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
mtcnn==0.0.9
numba==0.45.1
numpy==1.16.4
pandas==0.25.1
patsy==0.5.1
protobuf==3.9.1
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
scikit-learn==0.21.2
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
statsmodels==0.10.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.3
umap-learn==0.3.10
Werkzeug==0.15.5
wincertstore==0.2
wrapt==1.11.2
